# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.createCommentRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
  model.Comment:
    properties:
      author:
        $ref: '#/definitions/model.User'
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_hidden:
        type: boolean
      post_id:
        type: string
      status:
        $ref: '#/definitions/model.ContentStatus'
      updated_at:
        type: string
      user_id:
        type: string
      user_vote:
        type: integer
      vote_count:
        description: Champs calculés (non stockés en DB)
        type: integer
    type: object
  model.ContentStatus:
    enum:
    - pending_moderation
    - approved
    - flagged
    - blocked
    type: string
    x-enum-varnames:
    - StatusPendingModeration
    - StatusApproved
    - StatusFlagged
    - StatusBlocked
...
  model.User:
    properties:
      avatar_url:
        type: string
      banner_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      role:
        $ref: '#/definitions/model.UserRole'
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.UserRole:
    enum:
    - user
    - moderator
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleModerator
    - RoleAdmin
...
host: localhost:8080
info:
  contact:
    name: DevHelp Team
    url: https://github.com/bloiss/DevHelp
  description: API REST du forum DevHelp — entraide Dev & IA.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DevHelp API
  version: "1.0"
paths:
...
  /api/v1/posts/{id}/comments:
    get:
      description: Retourne tous les commentaires associés à un post, avec le vote
        de l'utilisateur connecté si disponible.
      parameters:
      - description: UUID post
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Comment'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lister les commentaires d'un post
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Ajoute un commentaire sur un post. L'auteur est déduit du token
        JWT.
      parameters:
      - description: UUID post
        in: path
        name: id
        required: true
        type: string
      - description: Contenu
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Créer un commentaire
      tags:
      - comments
  /api/v1/posts/{id}/comments/{commentId}:
    delete:
      description: Supprime un commentaire. Seul l'auteur du commentaire ou un admin
        peut supprimer.
      parameters:
      - description: UUID post
        in: path
        name: id
        required: true
        type: string
      - description: UUID commentaire
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Supprimer un commentaire
      tags:
      - comments
...
securityDefinitions:
  BearerAuth:
    description: Entrez "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"